Overhaul settings management
This moves the registration of default settings over to modules
themselves instead of subscreens, removing all the duplication, and make
the screens go through their respective module when accessing settings.
Most of the getters are now just a passthrough from a screen to the
abstract class, to its module, to the abstract class.

This fixes a crash where opening the General settings would crash,
as it would only initialize its values but tried to access stuff in the
Misc module.